Split JS files between src and test directories
