+++ bin/wasm-node/javascript/.npmignore
@@ -1,7 +1,3 @@
-demo.js
 prepare.js
-rococo_specs.js
 smoldot-demo-db.json
-test.js
-tick_specs.js
-westend_specs.js
+test
