+++ bin/wasm-node/javascript/package.json
@@ -13,14 +13,14 @@
     "url": "https://github.com/paritytech/smoldot/issues"
   },
   "type": "module",
-  "types": "index.d.ts",
-  "main": "index.js",
+  "types": "src/index.d.ts",
+  "main": "src/index.js",
   "scripts": {
     "prepublishOnly": "node prepare.js",
     "prestart": "node prepare.js",
     "build": "node prepare.js",
-    "start": "node demo.js",
-    "test": "node prepare.js && dtslint && node test.js"
+    "start": "node test/demo.js",
+    "test": "node prepare.js && dtslint && node test/test.js"
   },
   "browser": {
     "./compat-nodejs.js": "./compat-browser.js"
