+++ bin/wasm-node/javascript/prepare.js
@@ -33,14 +33,14 @@ child_process.execSync(
     { 'stdio': 'inherit' }
 );
 
-// It is then picked up by `wasm-opt`, which optimizes it and generates `./autogen/tmp.wasm`.
+// It is then picked up by `wasm-opt`, which optimizes it and generates `./src/autogen/tmp.wasm`.
 // `wasm_opt` is purely about optimizing. If it isn't available, it is also possible to directly
 // use the `.wasm` generated by the Rust compiler.
 let fallback_copy = false;
 try {
     if (build_profile == 'release') {
         child_process.execSync(
-            "wasm-opt -o autogen/tmp.wasm -Os --strip-debug --vacuum --dce "
+            "wasm-opt -o src/autogen/tmp.wasm -Os --strip-debug --vacuum --dce "
             + "../../../target/wasm32-wasi/" + build_profile + "/smoldot_js.wasm",
             { 'stdio': 'inherit' }
         );
@@ -53,15 +53,15 @@ try {
     fallback_copy = true;
 }
 if (fallback_copy) {
-    fs.copyFileSync("../../../target/wasm32-wasi/" + build_profile + "/smoldot_js.wasm", "./autogen/tmp.wasm");
+    fs.copyFileSync("../../../target/wasm32-wasi/" + build_profile + "/smoldot_js.wasm", "./src/autogen/tmp.wasm");
 }
 
 // We then base64-encode the `.wasm` file, and put this base64 string as a constant in
-// `./autogen/wasm.js`. It will be decoded at runtime.
-let wasm_opt_out = fs.readFileSync('./autogen/tmp.wasm');
+// `./src/autogen/wasm.js`. It will be decoded at runtime.
+let wasm_opt_out = fs.readFileSync('./src/autogen/tmp.wasm');
 let base64_data = wasm_opt_out.toString('base64');
-fs.writeFileSync('./autogen/wasm.js', 'export default "' + base64_data + '";');
-fs.unlinkSync("./autogen/tmp.wasm");
+fs.writeFileSync('./src/autogen/wasm.js', 'export default "' + base64_data + '";');
+fs.unlinkSync("./src/autogen/tmp.wasm");
 
 // The reason for this script is that at the time of writing, there isn't any standard
 // cross-platform solution to the problem of importing WebAssembly files. Base64-encoding the
