+++ bin/wasm-node/javascript/src/autogen/.gitignore
