+++ bin/wasm-node/javascript/src/autogen/.npmignore
