+++ bin/wasm-node/javascript/src/bindings-smoldot-js.js
