+++ bin/wasm-node/javascript/src/bindings-wasi.js
