+++ bin/wasm-node/javascript/src/compat-browser.js
