+++ bin/wasm-node/javascript/src/compat-nodejs.js
