+++ bin/wasm-node/javascript/src/index.d.ts
@@ -14,6 +14,7 @@
 
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
 declare class SmoldotError extends Error {
   constructor(message: string);
 }
