+++ bin/wasm-node/javascript/src/index.js
