+++ bin/wasm-node/javascript/src/index.test-d.ts
