+++ bin/wasm-node/javascript/src/worker.js
