+++ bin/wasm-node/javascript/test/test.js
