+++ bin/wasm-node/javascript/tsconfig.json
@@ -7,7 +7,7 @@
         "strictNullChecks": true,
         "strictFunctionTypes": true,
         "noEmit": true,
-        "baseUrl": ".",
+        "baseUrl": "./src",
         "paths": { "smoldot": ["."] }
     }
 }
